+++ models/post.go
@@ -7,10 +7,10 @@ type Post struct {
 	Author   string    `json:"author"`
 	Created  time.Time `json:"created"`
 	Forum    string    `json:"forum"`
-	ID       int64     `json:"id"`
+	ID       int       `json:"id"`
 	IsEdited bool      `json:"isEdited"`
 	Message  string    `json:"message"`
-	Parent   int64     `json:"parent"`
+	Parent   int       `json:"parent"`
 	Thread   int       `json:"thread"`
 	Path     string    `json:"-"`
 }
