+++ models/status.go
@@ -2,8 +2,8 @@ package models
 
 //Status - struct of status
 type Status struct {
-	Forum  int32 `json:"forum"`
-	Post   int64 `json:"post"`
-	Thread int32 `json:"thread"`
-	User   int32 `json:"user"`
+	Forum  int `json:"forum"`
+	Post   int `json:"post"`
+	Thread int `json:"thread"`
+	User   int `json:"user"`
 }
